+++ doc/source/doxygen-docs/changelog.md
@@ -8,6 +8,7 @@
     - API simplified: replace custom `mrpt::obs::gnss::gnss_message_ptr` with a `std::shared_ptr<>`.
 - BUG FIXES:
   - Fix wrong encoding of grayscale images in mrpt::ros1bridge::toROS() and mrpt::ros2bridge::toROS().
+  - Correctly return true/false in conversion of GPS observation in mrpt::ros1bridge::toROS() and mrpt::ros2bridge::toROS() depending on whether there is a valid GGA message.
 
 # Version 2.11.11: Released March 5th, 2024
 - Changes in libraries: